Vulkan: Move scissor/viewport rotation into the VulkanRenderManager. Fixes #12303.
