+++ ext/native/base/display.cpp
@@ -1,3 +1,4 @@
+#include <algorithm>
 #include "base/display.h"
 
 int dp_xres;
@@ -17,3 +18,36 @@ float display_hz = 60.0f;
 
 DisplayRotation g_display_rotation;
 Matrix4x4 g_display_rot_matrix;
+
+void RotateRectToDisplay(FRect &rect, float curRTWidth, float curRTHeight) {
+	switch (g_display_rotation) {
+	case DisplayRotation::ROTATE_180:
+		rect.x = curRTWidth - rect.w - rect.x;
+		rect.y = curRTHeight - rect.h - rect.y;
+		break;
+	case DisplayRotation::ROTATE_90: {
+		// Note that curRTWidth_ and curRTHeight_ are "swapped"!
+		float origX = rect.x;
+		float origY = rect.y;
+		float rtw = curRTHeight;
+		float rth = curRTWidth;
+		rect.x = rth - rect.h - origY;
+		rect.y = origX;
+		std::swap(rect.w, rect.h);
+		break;
+	}
+	case DisplayRotation::ROTATE_270: {
+		float origX = rect.x;
+		float origY = rect.y;
+		float rtw = curRTHeight;
+		float rth = curRTWidth;
+		rect.x = origY;
+		rect.y = rtw - rect.w - origX;
+		std::swap(rect.w, rect.h);
+		break;
+	}
+	case DisplayRotation::ROTATE_0:
+	default:
+		break;
+	}
+}
