+++ ext/native/base/display.h
@@ -30,3 +30,8 @@ enum class DisplayRotation {
 
 extern DisplayRotation g_display_rotation;
 extern Matrix4x4 g_display_rot_matrix;
+
+struct FRect {
+	float x, y, w, h;
+};
+void RotateRectToDisplay(FRect &rect, float curRTWidth, float curRTHeight);
