+++ ext/native/thin3d/VulkanRenderManager.h
@@ -10,6 +10,7 @@
 #include <mutex>
 #include <thread>
 
+#include "base/display.h"
 #include "Common/Vulkan/VulkanContext.h"
 #include "math/dataconv.h"
 #include "math/math_util.h"
@@ -119,13 +120,18 @@ public:
 	void SetViewport(const VkViewport &vp) {
 		_dbg_assert_(G3D, curRenderStep_ && curRenderStep_->stepType == VKRStepType::RENDER);
 		_dbg_assert_(G3D, (int)vp.width >= 0);
+		FRect rc{ vp.x, vp.y, vp.width, vp.height };
+		if (curRenderStep_->render.framebuffer == nullptr) { // Only the backbuffer is actually rotated wrong!
+			RotateRectToDisplay(rc, (float)vulkan_->GetBackbufferWidth(), (float)vulkan_->GetBackbufferHeight());
+		}
 		VkRenderData data{ VKRRenderCommand::VIEWPORT };
-		data.viewport.vp.x = vp.x;
-		data.viewport.vp.y = vp.y;
-		data.viewport.vp.width = vp.width;
-		data.viewport.vp.height = vp.height;
+		data.viewport.vp.x = rc.x;
+		data.viewport.vp.y = rc.y;
+		data.viewport.vp.width = rc.w;
+		data.viewport.vp.height = rc.h;
 		// We can't allow values outside this range unless we use VK_EXT_depth_range_unrestricted.
 		// Sometimes state mapping produces 65536/65535 which is slightly outside.
+		// TODO: This should be fixed at the source.
 		data.viewport.vp.maxDepth = clamp_value(vp.maxDepth, 0.0f, 1.0f);
 		data.viewport.vp.minDepth = clamp_value(vp.minDepth, 0.0f, 1.0f);
 		curRenderStep_->commands.push_back(data);
@@ -136,7 +142,15 @@ public:
 		_dbg_assert_(G3D, (int)rc.extent.width >= 0);
 		_dbg_assert_(G3D, (int)rc.extent.height >= 0);
 		VkRenderData data{ VKRRenderCommand::SCISSOR };
-		data.scissor.scissor = rc;
+		if (curRenderStep_->render.framebuffer == nullptr) {
+			FRect frc{ (float)rc.offset.x, (float)rc.offset.y, (float)rc.extent.width, (float)rc.extent.height };
+			if (curRenderStep_->render.framebuffer == nullptr) { // Only the backbuffer is actually rotated wrong!
+				RotateRectToDisplay(frc, (float)vulkan_->GetBackbufferWidth(), (float)vulkan_->GetBackbufferHeight());
+			}
+			data.scissor.scissor = VkRect2D{ { (int32_t)frc.x, (int32_t)frc.y }, { (uint32_t)frc.w, (uint32_t)frc.h} };
+		} else {
+			data.scissor.scissor = rc;
+		}
 		curRenderStep_->commands.push_back(data);
 	}
