+++ ext/native/thin3d/thin3d_vulkan.cpp
@@ -1059,13 +1059,7 @@ Pipeline *VKContext::CreateGraphicsPipeline(const PipelineDesc &desc) {
 }
 
 void VKContext::SetScissorRect(int left, int top, int width, int height) {
-	FRect rc{ (float)left, (float)top, (float)width, (float)height };
-	if (curFramebuffer_ == nullptr) { // Only the backbuffer is actually rotated wrong!
-		int curRTWidth, curRTHeight;
-		GetFramebufferDimensions((Framebuffer *)curFramebuffer_, &curRTWidth, &curRTHeight);
-		RotateRectToDisplay(rc, (float)curRTWidth, (float)curRTHeight);
-	}
-	VkRect2D scissor{ {(int32_t)rc.x, (int32_t)rc.y}, {(uint32_t)rc.w, (uint32_t)rc.h} };
+	VkRect2D scissor{ {(int32_t)left, (int32_t)top}, {(uint32_t)width, (uint32_t)height} };
 	renderManager_.SetScissor(scissor);
 }
 
@@ -1074,11 +1068,6 @@ void VKContext::SetViewports(int count, Viewport *viewports) {
 		// Ignore viewports more than the first.
 		VkViewport viewport;
 		FRect rc{ viewports[0].TopLeftX , viewports[0].TopLeftY, viewports[0].Width, viewports[0].Height };
-		if (curFramebuffer_ == nullptr) { // Only the backbuffer is actually rotated wrong!
-			int curRTWidth, curRTHeight;
-			GetFramebufferDimensions((Framebuffer *)curFramebuffer_, &curRTWidth, &curRTHeight);
-			RotateRectToDisplay(rc, (float)curRTWidth, (float)curRTHeight);
-		}
 		viewport.x = rc.x;
 		viewport.y = rc.y;
 		viewport.width = rc.w;
